hardware: clock/diff/reset: split in three files
